multi auth

+++ routes/web.php
@@ -19,20 +19,35 @@ Route::get('/register', function () {
     return view('register');
 });
 
-Route::get('/admin', 'AdminController@index')->name('admin');
+Route::get('/admin', 'AdminController@index')->name('admin')->middleware('auth:teachers');
 
-//auth
-Route::get('/login', 'AuthController@login');
-Route::post('/login', 'AuthController@postlogin')->name('login');
+Route::get('/siswa', 'SiswaController@index')->name('siswa')->middleware('auth:students');
+
+//auth siswa
+Route::get('/login', 'AuthController@login')->middleware('guest');
+Route::post('/loginn', 'AuthController@postlogin')->name('login');
 Route::get('/register-siswa', 'AuthController@registersiswa')->name('registersiswa');
-Route::post('/register-siswa', 'AuthController@postregister')->name('postregistersiswa');
+Route::post('/register-siswa', 'AuthController@postregistersiswa')->name('postregistersiswa');
+
+//auth guru
+Route::get('/register-guru', 'AuthController@registerguru')->name('registerguru');
+Route::post('/register-guru', 'AuthController@postregisterguru')->name('postregisterguru');
 Route::get('/logout', 'AuthController@logout')->name('logout');
 
 //siswa
-Route::get('/data-siswa', 'StudentsController@index')->name('data-siswa');
-Route::get('/detail-siswa/{student}', 'StudentsController@show');
-Route::get('/tambah-siswa/', 'StudentsController@create');
-Route::post('/data-siswa', 'StudentsController@store');
-Route::delete('/data-siswa/{student}', 'StudentsController@destroy');
-Route::get('/data-siswa/{student}/edit-siswa', 'StudentsController@edit');
-Route::patch('/data-siswa/{student}', 'StudentsController@update');
+Route::get('/data-siswa', 'StudentsController@index')->name('data-siswa')->middleware('auth:teachers');
+Route::get('/detail-siswa/{student}', 'StudentsController@show')->middleware('auth:teachers');
+Route::get('/tambah-siswa/', 'StudentsController@create')->middleware('auth:teachers');
+Route::post('/data-siswa', 'StudentsController@store')->middleware('auth:teachers');
+Route::delete('/data-siswa/{student}', 'StudentsController@destroy')->middleware('auth:teachers');
+Route::get('/data-siswa/{student}/edit-siswa', 'StudentsController@edit')->middleware('auth:teachers');
+Route::patch('/data-siswa/{student}', 'StudentsController@update')->middleware('auth:teachers');
+
+//guru
+Route::get('/data-guru', 'TeachersController@index')->name('data-guru')->middleware('auth:teachers');
+Route::get('/detail-guru/{teacher}', 'TeachersController@show')->middleware('auth:teachers');
+Route::get('/tambah-guru/', 'TeachersController@create')->middleware('auth:teachers');
+Route::post('/data-guru', 'TeachersController@store')->middleware('auth:teachers');
+Route::delete('/data-guru/{teacher}', 'TeachersController@destroy')->middleware('auth:teachers');
+Route::get('/data-guru/{teacher}/edit-guru', 'TeachersController@edit')->middleware('auth:teachers');
+Route::patch('/data-guru/{teacher}', 'TeachersController@update')->middleware('auth:teachers');
